add basic building
