+++ .gitignore
@@ -1,1 +1,2 @@
 package.json.js
+node_modules/
