+++ index.js
@@ -5,3 +5,4 @@
  */
 'use strict';
 
+window.CanvasZH = require('../lib/canvas_zh.js');
