+++ scripts/build.js
@@ -1,0 +1,16 @@
+/**
+ * Build the code for browser using
+ *
+ * @date 17/4/10
+ */
+'use strict';
+
+const path = require('path');
+const fs = require('fs');
+const browserify = require('browserify');
+
+browserify({
+    entry: path.join(__dirname, '../index.js')
+})
+.bundle()
+.pipe(fs.createWriteStream(path.join(__dirname, '../bin/canvas-zhihu.js')));
